requirements.txt
tweepy
python-dotenv
openai==0.28
pyyaml
requests